% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCorrelationMatrix.R
\name{getCorrelationMatrix}
\alias{getCorrelationMatrix}
\title{Produce CpG correlation matrix}
\usage{
getCorrelationMatrix(chr, data_cases_controls, r_low = 0.1,
  r_high = 0.3, dist_cutoff = 1e+08, one_prob_per_CoRSIV = F)
}
\arguments{
\item{m}{m is a data frame with columns are probe data and rows are samples}
}
\value{
returns a dataframe
}
\description{
This function takes input as dataframe with columns of probe data and rows are samples
and produce a correlation matrix
}
\examples{
m2beta(m)
}
\keyword{betavalues}
\keyword{mvalues}
