% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2beta.R
\name{m2beta}
\alias{m2beta}
\title{Convert m values to beta values}
\usage{
m2beta(m)
}
\arguments{
\item{m}{m is a data frame with columns are probe data and rows are samples}
}
\value{
returns a dataframe
}
\description{
This function takes input as dataframe with columns of probe data and rows are samples
}
\examples{
m2beta(m)
}
\keyword{betavalues}
\keyword{mvalues}
