% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{probe_id}
\alias{probe_id}
\title{HH450k probes overlapping CoRSIV regions}
\format{An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 3902 rows and 5 columns.}
\usage{
probe_id
}
\description{
HH450k probes overlapping CoRSIV regions
}
\examples{
\dontrun{
 probe_id
}
}
\keyword{datasets}
